requirements.txt
fastapi
uvicorn
python-dotenv
langgraph==0.6.4
langgraph-checkpoint-sqlite==2.0.11
langchain-openai==0.3.30
langchain-tavily==0.2.11
langchain-chroma==0.2.2
langchain-community==0.3.27
openai       
langchain    
qdrant-client
python-dotenv
bs4==0.0.2
PyMuPDF
python-docx
python-pptx
psycopg2-binary==2.9.10
